fix: rename macd.py to macd_strategy.py and update imports to match naming convention
