+++ tests/test_fuzz_strategies.py
@@ -4,7 +4,7 @@ import logging
 import pytest
 from trading_bot.strategies.sma_strategy import sma_strategy
 from trading_bot.strategies.rsi_strategy import rsi_strategy
-from trading_bot.strategies.macd_strategy import macd_strategy
+from trading_bot.strategies.macd_strategy_strategy import macd_strategy
 from trading_bot.strategies.bbands_strategy import bbands_strategy
