+++ trading_bot/strategy.py
@@ -1,6 +1,6 @@
 from trading_bot.strategies.sma_strategy import sma_strategy
 from trading_bot.strategies.rsi_strategy import rsi_strategy
-from trading_bot.strategies.macd_strategy import macd_strategy
+from trading_bot.strategies.macd_strategy_strategy import macd_strategy
 from trading_bot.strategies.bbands_strategy import bbands_strategy
 from trading_bot.strategies.confluence_strategy import confluence_strategy
